+++ client/src/app/modules/pixlisecore/models/expression-data-source.ts
@@ -11,7 +11,7 @@ import { QuantGetReq, QuantGetResp } from "src/app/generated-protos/quantificati
 import { ScanBeamLocationsReq, ScanBeamLocationsResp } from "src/app/generated-protos/scan-beam-location-msgs";
 import { ScanEntryReq, ScanEntryResp } from "src/app/generated-protos/scan-entry-msgs";
 import { ScanEntryMetadataReq, ScanEntryMetadataResp } from "src/app/generated-protos/scan-entry-metadata-msgs";
-import { SpectrumReq, SpectrumResp } from "src/app/generated-protos/spectrum-msgs";
+import { SpectrumResp } from "src/app/generated-protos/spectrum-msgs";
 import { PseudoIntensityReq, PseudoIntensityResp } from "src/app/generated-protos/pseudo-intensities-msgs";
 import { ScanEntryRange, ScanMetaDataType } from "src/app/generated-protos/scan";
 import { RegionOfInterestGetReq, RegionOfInterestGetResp } from "src/app/generated-protos/roi-msgs";
@@ -924,6 +924,10 @@ export class ExpressionDataSource
             const entry = scanMetaData.entries[c];
             if (entry) {
               const item = entry.meta[metaIdx];
+              if (item === undefined) {
+                throw new Error(`Scan entry ${c} does not contain housekeeping data labelled: "${name}"`);
+              }
+
               const value = (metaType == ScanMetaDataType.MT_FLOAT ? item.fvalue : item.ivalue) || 0; // Shut up compiler...
 
               const locIdx = this._scanEntryIndexesRequested[c];