+++ client/src/app/modules/scatterplots/widgets/rgbu-plot-widget/rgbu-plot-widget.component.ts
@@ -352,7 +352,12 @@ export class RGBUPlotWidgetComponent extends BaseWidgetModel implements OnInit, 
   }
 
   onSelectionClear(): void {
-    this._selectionService.clearSelection();
+    const allScanIds = [];
+    for (const scan of Object.values(this._analysisLayoutService.activeScreenConfiguration$.value.scanConfigurations)) {
+      allScanIds.push(scan.id);
+    }
+
+    this._selectionService.clearSelection(allScanIds);
   }
 
   onImageChanged(imagePath: string) {